--- src/main/kotlin/me/hydos/rosella/render/shader/pushconstant/ModelPushConstant.kt
@@ -1,7 +1,0 @@
-package me.hydos.rosella.render.shader.pushconstant
-
-import org.joml.Vector3f
-
-class ModelPushConstant {
-    val position: Vector3f = Vector3f()
-}
